refactor(analysis): split cache key/provenance and manifest scope walk
